Added running beam, do not use UrbanSim stuff in case it is disable, BEAM is relative to the BEAM local input folder
